+++ README.md
@@ -1,0 +1,13 @@
+Project Structure
+.
+├── index.html
+├── README.md
+└── src
+    ├── animationData.js
+    ├── classes
+    │   ├── player.class.js
+    │   └── uiManager.class.js
+    ├── main.js
+    ├── sprites
+    │   ├── ...
+    └── style.css
