--- guile/frames.js
@@ -1,2 +1,0 @@
-export const frameX = [0,1,2,3,4,5,0,1,2,3,4,5,0,1,2,3,4,5,6.2,7.3,8.5,9.7,5,6.2,7.3,8.5,9.7,5,6.2,7.3,8.5,9.7,10.7,11.8,10.7,11.8,10.7,11.8,12.9,13,14.4,15.5,16.7,17.9,19.1,20,21.3,22.5,23.8,24.9,26.1,27.3,28.6,29.7,30.9,32.1,34]
-export const frameY = [0];
