--- guile/index.html
@@ -1,15 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>First Animation</title>
-    <link rel="stylesheet" href="style.css">
-</head>
-<body>
-    <h1>Guile - Sonic Boom</h1>
-    <canvas id="canvas1"></canvas>
-
-    <script type="module" src="script.js"></script> 
-</body>
-</html>
