--- guile/style.css
@@ -1,15 +1,0 @@
-body {
-    text-align: center;
-    color: white;
-    background-color: #0d262d;
-}
-
-#canvas1 {
-    border: 5px solid black;
-    position: absolute;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    width: 400px;
-    height: 400px;
-}
