+++ index.html
@@ -5,9 +5,8 @@
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>Document</title>
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="./src/style.css">
 </head>
-
 <body>
     <canvas id="canvas1"></canvas>
 
@@ -27,7 +26,7 @@
         </select>
     </div>
 
-    <script src="script.js"></script>
+    <script type="module" src="./src/main.js"></script>
 </body>
 
 </html>
