--- note.txt
@@ -1,3 +1,0 @@
-33.03
-
-https://www.youtube.com/watch?v=CY0HE277IBM&t=2434s
