+++ src/animationData.js
@@ -1,0 +1,12 @@
+export const animationData = [
+    { name: 'idle', frames: 5 },
+    { name: 'run', frames: 8 },
+    { name: 'hurt', frames: 2 },
+    { name: 'dead', frames: 5 },
+    { name: 'jump', frames: 8 },
+    { name: 'attack1', frames: 6 },
+    { name: 'attack2', frames: 3 },
+    { name: 'attack3', frames: 3 },
+    { name: 'attack4', frames: 10 },
+    { name: 'walk', frames: 8 }
+];
