+++ src/sprites/dead.png
