+++ src/sprites/hurt.png
